+++ noodles-csi/src/io.rs
@@ -1,11 +1,12 @@
 //! CSI I/O.
 
 mod filter_by_region;
+mod indexed_reader;
 mod indexed_record;
 mod indexed_records;
 mod query;
 
 pub use self::{
-    filter_by_region::FilterByRegion, indexed_record::IndexedRecord,
+    filter_by_region::FilterByRegion, indexed_reader::IndexedReader, indexed_record::IndexedRecord,
     indexed_records::IndexedRecords, query::Query,
 };