+++ noodles-tabix/examples/tabix_query_generic.rs
@@ -4,24 +4,10 @@
 //!
 //! The result matches the output of `tabix <src> <region>`.
 
-use std::{env, fs::File, io};
+use std::{env, fs::File};
 
-use noodles_bgzf as bgzf;
-use noodles_core::Region;
 use noodles_csi as csi;
 use noodles_tabix as tabix;
-
-fn resolve_region(header: &csi::index::Header, region: &Region) -> io::Result<usize> {
-    header
-        .reference_sequence_names()
-        .get_index_of(region.name())
-        .ok_or_else(|| {
-            io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "missing reference sequence name",
-            )
-        })
-}
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut args = env::args().skip(1);
@@ -30,19 +16,11 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let tabix_src = format!("{src}.tbi");
     let index = tabix::read(tabix_src)?;
-    let header = index
-        .header()
-        .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidInput, "missing tabix header"))?;
+    let mut reader = File::open(src).map(|f| csi::io::IndexedReader::new(f, index))?;
 
-    let mut reader = File::open(src).map(bgzf::Reader::new)?;
+    let query = reader.query(&region)?;
 
-    let reference_sequence_id = resolve_region(header, &region)?;
-    let chunks = index.query(reference_sequence_id, region.interval())?;
-    let records = csi::io::Query::new(&mut reader, chunks)
-        .indexed_records(header)
-        .filter_by_region(&region);
-
-    for result in records {
+    for result in query {
         let record = result?;
         println!("{}", record.as_ref());
     }